refactor: remove unused MarketFaq and change size
