--- apps/main/src/llamalend/features/market-faq/index.ts
@@ -1,1 +1,0 @@
-export { MarketFaq, MarketFaqCard } from './MarketFaq'
